Rename utils/format.py to utils/format_trials.py
